% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{create_poly}
\alias{create_poly}
\title{Create polygon from a list or vector of longitude and latitude
coordinates}
\usage{
create_poly(longitude, latitude, crs = 4326)
}
\arguments{
\item{longitude}{Vector or list with longitude coordinates}

\item{latitude}{Vector or list with latitude coordinates}

\item{crs}{number or string representing a coordinate reference system (CRS)}
}
\value{
Polygon as sfc class defined in "sf" package.
}
\description{
Create polygon as class "sfc" form sf R package by provinding
a list or vector of longitude and latitude. If the coordinates are projected
on a defined crs. The crs can be provided to properly construct the polygon
via the crs argument. If only two coordinates are given, a rectangle polygon
is created using the two points as top left and right bottom corner.
The coordinates doesn't need to close, the function take care of it for you.
}
\examples{
polygone = create_poly(c(-70,-60),c(40,45))
polygone = create_poly(c(-65,-70,-70,-65),c(40,40,45,45), crs = 4326)
}
