% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_functions.R
\name{doc_nc}
\alias{doc_nc}
\title{Print global attributes of a NetCDF file}
\usage{
doc_nc(varname, path_to_data = NULL)
}
\arguments{
\item{varname}{Name of the variable whose source file will be inspected.}

\item{path_to_data}{Default = NULL. Path to the data directory.
If NULL, the path stored in the package internal file is used.}
}
\value{
Invisibly returns the named list of global attributes.
}
\description{
Find the NetCDF file containing the given variable and
pretty-print its global attributes.
}
\examples{
\dontrun{doc_nc("bottom_temperature")}

}
