% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{indistance_var}
\alias{indistance_var}
\title{Select the data within a distance of a point}
\usage{
indistance_var(var_list, lon, lat, dist = 5000)
}
\arguments{
\item{var_list}{list containing the variables extracted from \code{\link{read_nc}} or \code{\link{load_var}}}

\item{lon}{longitude coordinate of the point}

\item{lat}{latitude coordinate of the point}

\item{dist}{distance, in meter, of the coordinates to keep the data.
default is 5000 meters}
}
\value{
var_list containing the variables, coordinates and metadata of the selected data
}
\description{
Select the data that is located within a certain distance of a
specified geographic coordinate.
}
\examples{
\dontrun{masked_list = indistance_polygon(var_list, lon=-65, lat=49, dist=15000)}
}
