% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{inpolygon_var}
\alias{inpolygon_var}
\title{Mask data that are outside a chosen polygon}
\usage{
inpolygon_var(var_list, pol)
}
\arguments{
\item{var_list}{list containing the variables extracted from \code{\link{read_nc}} or \code{\link{load_var}}}

\item{pol}{polygon of class "sf", "sfc" or "sfg"}
}
\value{
var_list containing the variables, coordinates and metadata of the selected data
}
\description{
Mask and slice the data of a var_list that are outside the polygon
}
\examples{
\dontrun{masked_list = inpolygon_var(var_list, polygone)}
}
