% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{inside_poly}
\alias{inside_poly}
\title{Mask data that are outside a chosen polygon}
\usage{
inside_poly(var_list, pol)
}
\arguments{
\item{var_list}{list containing the variables extracted from \code{\link{read_nc}} or \code{\link{load_var}}}

\item{pol}{polygon of class "sf" or "sfc"}
}
\value{
a list with data masked outside the polygon
}
\description{
Mask and slice the data of a var_list that are outside the polygon
}
\examples{
\dontrun{masked_list = inside_poly(bottom_temp, polygone)}
}
