% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_functions.R
\name{list_nc}
\alias{list_nc}
\title{List the variables in the data folder}
\usage{
list_nc()
}
\description{
Provide information on the variables found in the data folder set by
\code{\link{setDataPath}}
}
