% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{load_poly}
\alias{load_poly}
\title{Load a polygon from file}
\usage{
load_poly(load_path, lon_name = "longitude", lat_name = "latitude")
}
\arguments{
\item{load_path}{Path to load the file}

\item{lon_name}{name given to the columns with the longitude component in
the loaded file. Arguments only works for csv file}

\item{lat_name}{name given to the columns with the latitude component in
the loaded file. Arguments only works for csv file}
}
\value{
Polygon as sfc class defined in "sf" package is extension is .csv.
Return a list comprising a polygon of sfc class.
}
\description{
Load a a polygon from a file. For the moment, only .csv and .rdata
are implemented. The loading method is defined by the filename extension.
}
\examples{
\dontrun{polygone = load_poly("/home/bruno/<save directory>/save_poly.rdata")}

}
