% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_var.R
\name{load_var}
\alias{load_var}
\title{Load variable to a var_list}
\usage{
load_var(loadpath)
}
\arguments{
\item{loadpath}{Path to the file containing the data to load.}
}
\value{
variable data as a list containing the data, shape, dimensions
 (longitude ,latitude ,time) and units.
}
\description{
Function returning data as a list from saved file created
 by \code{\link{save_var}}
}
\examples{
\dontrun{data <- load_var("/home/bruno/workspace/Data/test/saved_data.rdata")}

}
