% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{nc_datetime2R_datetime}
\alias{nc_datetime2R_datetime}
\title{Convert various time format from a NetCDF file to an R datetime}
\usage{
nc_datetime2R_datetime(nc)
}
\arguments{
\item{nc}{A ncdf4 object}
}
\value{
Numeric Second since epoch time 1970-01-01
}
\description{
This function doubles a number. It is not exported.
}
\keyword{internal}
