% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_functions.R
\name{read_nc}
\alias{read_nc}
\title{Read and import environmental data.}
\usage{
read_nc(
  varname,
  longitude = NULL,
  latitude = NULL,
  time = NULL,
  path_to_data = NULL,
  date_format = "\%Y-\%m"
)
}
\arguments{
\item{varname}{Name of the variable to read}

\item{longitude}{Default = NA, Range of longitude to slice the variable data.
If NA, the whole dimensions is imported. The longitude argument must be NA or
a vector of length 2.}

\item{latitude}{Default = NA, Range of latitude to slice the variable data.
If NA, the whole dimension is imported. The latitude argument must be NA or
a vector of length 2.}

\item{time}{Default = NA, Range of time to slice the variable data.
If NA, the whole dimension is imported. The time argument must be NA or a
vector of length 2.}

\item{path_to_data}{Default = NA. The path to the data directory.
If NA, the path used is the default one stored in the package internal file.}

\item{date_format}{Default = "%Y-%m". The format of the time argument.
Default can be seen as "YYYY-MM".
Other possible format are available via \code{\link{strptime}}}
}
\value{
Return a list containing the sliced data, dimensions and units.

To print all the possible field names, use the command \code{\link{names}}.
}
\description{
This function read the NetCDF (.nc) file associated with the
environmental data, .
The function return a list containing the sliced data, its coordinates and
others useful metadata.
}
\examples{
\dontrun{var = read_nc("bottom_temperature")

var = read_nc("bottom_temperature", longitude=c(-64,-50), latitude=c(34,50),
time = c("1985-01","2004-11"))}

}
