% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{save_poly}
\alias{save_poly}
\title{Save polygon on local disk}
\usage{
save_poly(pol, savepath, polyname = NA)
}
\arguments{
\item{pol}{Polygon as sf | sfc class from the R package "sf"}

\item{savepath}{Path to save the polygone}

\item{polyname}{Name of the polygone, Default = NA, this arguments is only
taken into account if the extension of savepath is "rdata". In this case,
polyname must be a char type.}
}
\description{
Save polygon on the local computer in various format
depending o the extension of the savepath argument. For the moment, only
"rdata" and "csv" are implemented.
}
\examples{
\dontrun{save_poly(polgone,"/home/bruno/save_directory/save_poly.rdata",
polyname = "16F")}

}
