% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_var.R
\name{save_var}
\alias{save_var}
\title{save_var}
\usage{
save_var(var_list, savepath)
}
\arguments{
\item{var_list}{variable list created with this package}

\item{savepath}{Path to save the data. The format of the save file is
determined by the extension of this argument.}
}
\description{
Function that save a variable list created by read_nc or
load_var in different format.
}
\examples{
\dontrun{
save_var(data,"/home/bruno/test_data.nc",format = "nc")}

}
