% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Path.R
\name{setPolyPath}
\alias{setPolyPath}
\title{Set the path to the folder where the polygons are stored}
\usage{
setPolyPath(path_to_poly)
}
\arguments{
\item{path_to_poly}{Path to the default polygon folder}
}
\description{
Modifies the default path of the polygon folder in the
pkg_params.rdata file.
}
\examples{
\dontrun{setPolyPath("/home/bruno/workspace/Data/Polygon/")}
}
