% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{show_poly}
\alias{show_poly}
\title{Show polygon on the map}
\usage{
show_poly(pol)
}
\arguments{
\item{pol}{a polygone of either "sf", "sfc" class or any polygone
with crs attributes}
}
\description{
Show a polygon on the map of the estuary and the gulf of
St.Lawrence.
}
