% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_var.R
\name{stat_var}
\alias{stat_var}
\title{Perform statistical analysis on a var_list}
\usage{
stat_var(var_list, op, dims, prct = c(0.1, 0.9))
}
\arguments{
\item{var_list}{list containing the variable extracted from \code{\link{read_nc}} or \code{\link{load_var}}}

\item{op}{operations to perform on the data, character or list of characters.
Operations are "mean","min","max","med" for median, "sd" for standard
deviation, "prct" for percentile, "all" for all of the previous.}

\item{dims}{Output data dimensions, dims = "xy" or "t". If "xy", return a
spatial grid where the analysis is performed along time. If "t", return a
time series on which the analysis is performed along the longitude and
latitude dimensions for each time step.}

\item{prct}{Float or vector where element are between 0 and 1 to calculate
the percentile, 0.1 = 10%. Default = c(0.1,0.9). Only works if "prct" or
"all" is include in op. argument.}
}
\value{
list containing the result from the statistical analysis, the unit,
the dimensions and the number of valid observations for each points
}
\description{
Perform one or multiple statistical analysis along the time or
spatial dimension.
}
\examples{
\dontrun{temp <- stat_var(var_list,c("mean","min"),"xy")
temp <- stat_var(var_list,"all", "t", prct = c(0.25,0.5,0.75))}
}
