% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{viz_ts}
\alias{viz_ts}
\title{Visualize time series}
\usage{
viz_ts(data, var_list)
}
\arguments{
\item{data}{vector, or 1D object}

\item{var_list}{var_list containing information about the data}
}
\description{
Function to visualize 1D data along time.
}
\examples{
\dontrun{viz_ts(var_list$bottom_temperature, var_list)}
}
