% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_functions.R
\name{within_dist}
\alias{within_dist}
\title{Select the data inside a distance of a point}
\usage{
within_dist(var_list, lon, lat, dist = 5000)
}
\arguments{
\item{var_list}{list containing the variables extracted from \code{\link{read_nc}} or \code{\link{load_var}}}

\item{lon}{longitude coordinate of the point}

\item{lat}{latitude coordinate of the point}

\item{dist}{distance in meter of the center of the point}
}
\value{
var_list
}
\description{
Select the data that is located inside a specified distance of a
selected geographic coordinate.
}
\examples{
\dontrun{masked_list = within_dist(bottom_temp, lon=-65, lat=49, dist=5000)}
}
