% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_functions.R
\name{bymonth_var}
\alias{bymonth_var}
\title{Slice data from a variable list according to specified months}
\usage{
bymonth_var(var_list, month_num)
}
\arguments{
\item{var_list}{var_list created or loaded by this package}

\item{month_num}{Numeric number of the month (11 = november). Either numeric
or vector of numeric elements}
}
\value{
var_list containing data of the specified month
}
\description{
Extract data from a var_list by specifying which month the data
should contain.
}
\examples{
\dontrun{temp_june <- bymonth_var(total_temp,6)
temp_june_august <- bymonth_var(total_temp,c(06,08))}

}
