% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_functions.R
\name{info_var}
\alias{info_var}
\title{info_var}
\usage{
info_var(var_list)
}
\arguments{
\item{ls}{variable list created by this package}
}
\value{
The output from \code{\link{print}}
}
\description{
Print information about a variable list created by read_nc.
The variable name, shape, dimension names, units and memory used are printed.
This function is useful to quickly get information about a variable.
}
\examples{
\dontrun{var_info(ls)}
}
