% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_nc.R
\name{list_nc}
\alias{list_nc}
\title{List the variables in the data folder}
\usage{
list_nc(varname = NA)
}
\arguments{
\item{varname}{Variable to extract information about. Default = NA. If NA,
it print the available variable to be read by \code{\link{read_nc}}.
If varname is type character, information about the variable is print from
the associated netcdf file global attributes.}
}
\description{
Print the variables in the default data folder set by
\code{\link{setDataPath}}, or print information about the variable mentionned
as the varname argument.
}
