% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nc.R
\name{read_nc}
\alias{read_nc}
\title{Read netCDF file containing the gridded data.}
\usage{
read_nc(
  varname,
  longitude = NA,
  latitude = NA,
  time = NA,
  path_to_data = NA,
  date_format = "\%Y-\%m"
)
}
\arguments{
\item{varname}{Name of the variable}

\item{longitude}{Default = NA, Range of longitude to slice the variable data.
If NA, the whole dimensions is imported. The longitude argument must be NA or
a vector of length 2.}

\item{latitude}{Default = NA, Range of latitude to slice the variable data.
If NA, the whole dimension is imported. The latitude argument must be NA or
a vector of length 2.}

\item{time}{Default = NA, Range of time to slice the variable data.
If NA, the whole dimension is imported. The time argument must be NA or a
vector of length 2.}

\item{path_to_data}{Default = NA. The path to the data directory.
If NA, the path is store in /extdata/pkg_params.rdata under the name data_path}

\item{date_format}{Default = "%Y-%m". The format of the time argument.
Default can be seen as "YYYY-MM".
Other possible format are available via \code{\link{strptime}}}
}
\value{
Return a list containing the sliced data, its dimensions and their
values and the variable units.

To get all the possible field names, use the command \code{\link{names}}.
Example : names(var_data)
}
\description{
This function read the netCDF (.nc) file associated with the varname
argument, extract the data and slice it according to specified arguments.
The function return a list containing the sliced data, its coordinates and
others useful metadata.
}
\examples{
\dontrun{var = read_nc("bottom_temperature")

var = read_nc("bottom_temperature", longitude=c(-64,-50), latitude=c(34,50),
time = c("1985-01","2004-11"))}

}
