% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_functions.R
\name{save_var}
\alias{save_var}
\title{Save variable in a local file}
\usage{
save_var(var_list, savepath)
}
\arguments{
\item{var_list}{variable list created with this package}

\item{savepath}{Path to save the data. The format of the save file is
determined by the extension of this argument.}
}
\description{
Function that save a variable list created by \code{\link{read_nc}} or
\code{\link{load_var}} in csv or nc format.
}
\examples{
\dontrun{
save_var(var_list,"/home/bruno/test_data.nc"}

}
