% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Path.R
\name{setDataPath}
\alias{setDataPath}
\title{Set the default path to get the environmental data}
\usage{
setDataPath(path_to_data = NULL)
}
\arguments{
\item{path_to_data}{Path to the default data folder, Windows (\\) or Linux (/)
path style are supported even for Windows architecture, see example.
if NA, the default path is displayed.}
}
\description{
Tell this package where to find the environmental data.
}
\examples{
\dontrun{setDataPath("C:/Users/stdenisb/Documents/Coding/data/gslcoenv_data/")}

}
