% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Path.R
\name{setDataPath}
\alias{setDataPath}
\title{Set the path to the folder where the data are stored}
\usage{
setDataPath(path_to_data)
}
\arguments{
\item{path_to_data}{Path to the default data folder}
}
\description{
Modifies the default path of the data folder in the
pkg_params.rdata file.
}
\examples{
\dontrun{setDataPath("/home/bruno/workspace/Data/GSL-grid/")}

}
