% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_functions.R
\name{viz_map}
\alias{viz_map}
\title{Plot 2D data on a map ** Not fully implementend yet}
\usage{
viz_map(data, var_list, timeid = NA, cmap = cmocean::cmocean("thermal"))
}
\arguments{
\item{data}{Data to be plotted}

\item{var_list}{var_lsit containing informations about the data}

\item{timeid}{time index in case of 3D data, only works if length(dims(data)) > 2}

\item{cmap}{colormap used to plot the data, default = thermal from the
\code{\link[cmocean]{cmocean}} package.}
}
\description{
Plot 2D data on a map ** Not fully implementend yet
}
\examples{
\dontrun{viz_map(temp$data,temp,timeid=1)}
}
